pyproject.toml
[project]
name = "carcv"
version = "0.1.0"
description = "CarCV - Computer Vision system for vehicle recognition and license plate OCR"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "requests>=2.31.0",
    "python-Levenshtein>=0.21.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "pyarrow>=22.0.0",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "onnxruntime-gpu>=1.24.3",
    "opencv-python-headless>=4.8.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "polars>=1.40.0",
    "hydra-core>=1.3.2",
    "paramiko>=4.0.0",
    "pycocotools>=2.0.0",
    "pyyaml>=6.0",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }
